requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
djangorestframework==3.14.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
